+++ app/src/main/java/com/example/gamev3/Level0GameView.java
@@ -17,48 +17,48 @@ public class Level0GameView extends GameView {
     @Override
     public void setup() {
         commonSetup();
-        double height = 0.1;
-        double width = height * canvasHeight / canvasWidth;
-        double platformSizeX = 10 / canvasWidth;
-        double platformSizeY = 10 / canvasHeight;
-        player = new Player((float) 0, (float) 0.5,
+        float height = 0.1f;
+        float width = height * canvasHeight / canvasWidth;
+        float platformSizeX = 10 / canvasWidth;
+        float platformSizeY = 10 / canvasHeight;
+        player = new Player(0, 0.5f,
                 width, height);
 
         platforms = new ArrayList<>();
-        platforms.add(new Platform(0.2, 0.6, 0.1, platformSizeY));
+        platforms.add(new Platform(0.2f, 0.6f, 0.1f, platformSizeY));
 
-        platforms.add(new Platform(0.4, 0.4, 0.1, platformSizeY));
-        platforms.add(new Platform(0.6, 0.2, 0.1, platformSizeY));
-        platforms.add(new Platform(0.6, 0.6, 0.1, platformSizeY));
-        platforms.add(new Platform(1.0, 0.4, 0.15, platformSizeY));
-        platforms.add(new Platform(1.3, 0.4, 0.12, platformSizeY));
-        platforms.add(new Platform(1.55, 0.4, 0.1, platformSizeY));
-        platforms.add(new Platform(1.8, 0.3, 0.1, platformSizeY));
-        platforms.add(new Platform(2.0, 0.2, 0.1, platformSizeY));
+        platforms.add(new Platform(0.4f, 0.4f, 0.1f, platformSizeY));
+        platforms.add(new Platform(0.6f, 0.2f, 0.1f, platformSizeY));
+        platforms.add(new Platform(0.6f, 0.6f, 0.1f, platformSizeY));
+        platforms.add(new Platform(1.0f, 0.4f, 0.15f, platformSizeY));
+        platforms.add(new Platform(1.3f, 0.4f, 0.12f, platformSizeY));
+        platforms.add(new Platform(1.55f, 0.4f, 0.1f, platformSizeY));
+        platforms.add(new Platform(1.8f, 0.3f, 0.1f, platformSizeY));
+        platforms.add(new Platform(2.0f, 0.2f, 0.1f, platformSizeY));
 
 
-        platforms.add(new Platform(1. / 2, 0.9, 15, platformSizeY));
-        platforms.add(new Platform(-0.3, 0, platformSizeX, 1.8));
-        platforms.add(new Platform(2.3, 0, platformSizeX, 1.8));
+        platforms.add(new Platform(1.f / 2, 0.9f, 15, platformSizeY));
+        platforms.add(new Platform(-0.3f, 0, platformSizeX, 1.8f));
+        platforms.add(new Platform(2.3f, 0, platformSizeX, 1.8f));
 
         spikes = new ArrayList<>();
         boolean drawImageSpikes = gameProgress.graphicsLevel >= GameProgress.DANGER_LEVEL_FOR_SPIKES;
         if (gameProgress.dangerLevel >= GameProgress.DANGER_LEVEL_FOR_SPIKES) {
-            spikes.add(new Spike((float) 0.4, (float) 0.4, (float) 0.03, (float) 0.04, drawImageSpikes, spikeImage));
-            spikes.add(new Spike((float) 0.6, (float) 0.6, (float) 0.1, (float) 0.05, drawImageSpikes, spikeImage));
-            spikes.add(new Spike((float) 1.3, (float) 0.4, (float) 0.05, (float) 0.05, drawImageSpikes, spikeImage));
+            spikes.add(new Spike(0.4f, 0.4f, 0.03f, 0.04f, drawImageSpikes, spikeImage));
+            spikes.add(new Spike(0.6f, 0.6f, 0.1f, 0.05f, drawImageSpikes, spikeImage));
+            spikes.add(new Spike(1.3f, 0.4f, 0.05f, 0.05f, drawImageSpikes, spikeImage));
         }
 
         saws = new ArrayList<>();
         boolean drawImageSaw = gameProgress.graphicsLevel >= GameProgress.GRAPHICS_LEVEL_FOR_SAWS_IMAGES;
         if (gameProgress.dangerLevel >= GameProgress.DANGER_LEVEL_FOR_SAWS) {
-            double sawHeight = 0.1;
-            double sawWidth = sawHeight * canvasHeight / canvasWidth;
-            saws.add(new Saw((float) 0.4, (float) 0.4, sawWidth, sawHeight, 0.05, 1, drawImageSaw, sawImage));
-            saws.add(new Saw((float) 1.8, (float) 0.3, sawWidth, sawHeight, 0.05, 0.5, drawImageSaw, sawImage));
+            float sawHeight = 0.1f;
+            float sawWidth = sawHeight * canvasHeight / canvasWidth;
+            saws.add(new Saw(0.4f, 0.4f, sawWidth, sawHeight, 0.05f, 1f, drawImageSaw, sawImage));
+            saws.add(new Saw(1.8f, 0.3f, sawWidth, sawHeight, 0.05f, 0.5f, drawImageSaw, sawImage));
         }
 
 
-        finish = Finish.from_bottom_part(2, 0.2, canvasWidth, canvasHeight, finishImage, finishDoorImage);
+        finish = Finish.from_bottom_part(2, 0.2f, canvasWidth, canvasHeight, finishImage, finishDoorImage);
     }
 }
